Add files via upload
This was by far the hardest of the excersice. I learned to use dynamic arrays using vectors but for some reason my VScode just would not run the code even when it compiled so I had to use an online compiler, eventually I made it work and translated it to english.